pulp::aarch64

Struct Neon

pub struct Neon {
    pub neon: Neon,
}

Fields
neon: Neon
Implementations

impl Neon

pub unsafe fn new_unchecked() -> Self

Returns a SIMD token type without checking if the required CPU features for this type are available.

Safety

the required CPU features must be available.

pub fn try_new() -> Option<Self>

Returns a SIMD token type if the required CPU features for this type are available, otherwise returns None.

pub fn is_available() -> bool

Returns true if the required CPU features for this type are available, otherwise returns false.

pub fn vectorize<F: NullaryFnOnce>(self, f: F) -> F::Output

Vectorizes the given function as if the CPU features for this type were applied to it.

Note

For the vectorization to work properly, the given function must be inlined. Consider marking it as #[inline(always)]

pub fn to_ref(self) -> &'static Self

Takes a proof of the existence of this SIMD token (self), and returns a persistent reference to it.

impl Neon

pub fn add_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Adds the elements of each lane of a and b.

pub fn add_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Adds the elements of each lane of a and b.

pub fn and_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Returns the bitwise AND of a and b.

pub fn and_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Returns the bitwise AND of a and b.

pub fn and_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Returns the bitwise AND of a and b.

pub fn and_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Returns the bitwise AND of a and b.

pub fn and_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Returns the bitwise AND of a and b.

pub fn and_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Returns the bitwise AND of a and b.

pub fn and_m16x8(self, a: m16x8, b: m16x8) -> m16x8

Returns the bitwise AND of a and b.

pub fn and_m32x4(self, a: m32x4, b: m32x4) -> m32x4

Returns the bitwise AND of a and b.

pub fn and_m64x2(self, a: m64x2, b: m64x2) -> m64x2

Returns the bitwise AND of a and b.

pub fn and_m8x16(self, a: m8x16, b: m8x16) -> m8x16

Returns the bitwise AND of a and b.

pub fn and_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Returns the bitwise AND of a and b.

pub fn and_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Returns the bitwise AND of a and b.

pub fn and_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Returns the bitwise AND of a and b.

pub fn and_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Returns the bitwise AND of a and b.

pub fn andnot_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Returns the bitwise AND of NOT a and b.

pub fn andnot_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Returns the bitwise AND of NOT a and b.

pub fn andnot_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Returns the bitwise AND of NOT a and b.

pub fn andnot_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Returns the bitwise AND of NOT a and b.

pub fn andnot_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Returns the bitwise AND of NOT a and b.

pub fn andnot_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Returns the bitwise AND of NOT a and b.

pub fn andnot_m16x8(self, a: m16x8, b: m16x8) -> m16x8

Returns the bitwise AND of NOT a and b.

pub fn andnot_m32x4(self, a: m32x4, b: m32x4) -> m32x4

Returns the bitwise AND of NOT a and b.

pub fn andnot_m64x2(self, a: m64x2, b: m64x2) -> m64x2

Returns the bitwise AND of NOT a and b.

pub fn andnot_m8x16(self, a: m8x16, b: m8x16) -> m8x16

Returns the bitwise AND of NOT a and b.

pub fn andnot_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Returns the bitwise AND of NOT a and b.

pub fn andnot_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Returns the bitwise AND of NOT a and b.

pub fn andnot_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Returns the bitwise AND of NOT a and b.

pub fn andnot_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Returns the bitwise AND of NOT a and b.

pub fn cmp_eq_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_i16x8(self, a: i16x8, b: i16x8) -> m16x8

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_i32x4(self, a: i32x4, b: i32x4) -> m32x4

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_i64x2(self, a: i64x2, b: i64x2) -> m64x2

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_i8x16(self, a: i8x16, b: i8x16) -> m8x16

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_u16x8(self, a: u16x8, b: u16x8) -> m16x8

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_u32x4(self, a: u32x4, b: u32x4) -> m32x4

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_u64x2(self, a: u64x2, b: u64x2) -> m64x2

Compares the elements in each lane of a and b for equality.

pub fn cmp_eq_u8x16(self, a: u8x16, b: u8x16) -> m8x16

Compares the elements in each lane of a and b for equality.

pub fn cmp_ge_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_i16x8(self, a: i16x8, b: i16x8) -> m16x8

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_i32x4(self, a: i32x4, b: i32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_i64x2(self, a: i64x2, b: i64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_i8x16(self, a: i8x16, b: i8x16) -> m8x16

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_u16x8(self, a: u16x8, b: u16x8) -> m16x8

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_u32x4(self, a: u32x4, b: u32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_u64x2(self, a: u64x2, b: u64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_ge_u8x16(self, a: u8x16, b: u8x16) -> m8x16

Compares the elements in each lane of a and b for greater-than-or-equal-to.

pub fn cmp_gt_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_i16x8(self, a: i16x8, b: i16x8) -> m16x8

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_i32x4(self, a: i32x4, b: i32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_i64x2(self, a: i64x2, b: i64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_i8x16(self, a: i8x16, b: i8x16) -> m8x16

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_u16x8(self, a: u16x8, b: u16x8) -> m16x8

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_u32x4(self, a: u32x4, b: u32x4) -> m32x4

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_u64x2(self, a: u64x2, b: u64x2) -> m64x2

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_gt_u8x16(self, a: u8x16, b: u8x16) -> m8x16

Compares the elements in each lane of a and b for greater-than.

pub fn cmp_le_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_i16x8(self, a: i16x8, b: i16x8) -> m16x8

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_i32x4(self, a: i32x4, b: i32x4) -> m32x4

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_i64x2(self, a: i64x2, b: i64x2) -> m64x2

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_i8x16(self, a: i8x16, b: i8x16) -> m8x16

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_u16x8(self, a: u16x8, b: u16x8) -> m16x8

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_u32x4(self, a: u32x4, b: u32x4) -> m32x4

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_u64x2(self, a: u64x2, b: u64x2) -> m64x2

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_le_u8x16(self, a: u8x16, b: u8x16) -> m8x16

Compares the elements in each lane of a and b for less-than-or-equal-to.

pub fn cmp_lt_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_i16x8(self, a: i16x8, b: i16x8) -> m16x8

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_i32x4(self, a: i32x4, b: i32x4) -> m32x4

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_i64x2(self, a: i64x2, b: i64x2) -> m64x2

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_i8x16(self, a: i8x16, b: i8x16) -> m8x16

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_u16x8(self, a: u16x8, b: u16x8) -> m16x8

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_u32x4(self, a: u32x4, b: u32x4) -> m32x4

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_u64x2(self, a: u64x2, b: u64x2) -> m64x2

Compares the elements in each lane of a and b for less-than.

pub fn cmp_lt_u8x16(self, a: u8x16, b: u8x16) -> m8x16

Compares the elements in each lane of a and b for less-than.

pub fn cmp_not_eq_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for inequality.

pub fn cmp_not_eq_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for inequality.

pub fn cmp_not_ge_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for not-greater-than-or-equal.

pub fn cmp_not_ge_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for not-greater-than-or-equal.

pub fn cmp_not_gt_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for not-greater-than.

pub fn cmp_not_gt_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for not-greater-than.

pub fn cmp_not_le_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for not-less-than-or-equal.

pub fn cmp_not_le_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for not-less-than-or-equal.

pub fn cmp_not_lt_f32x4(self, a: f32x4, b: f32x4) -> m32x4

Compares the elements in each lane of a and b for not-less-than.

pub fn cmp_not_lt_f64x2(self, a: f64x2, b: f64x2) -> m64x2

Compares the elements in each lane of a and b for not-less-than.

pub fn div_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Divides the elements of each lane of a and b.

pub fn div_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Divides the elements of each lane of a and b.

pub fn is_nan_f32x4(self, a: f32x4) -> m32x4

Checks if the elements in each lane of a are NaN.

pub fn is_nan_f64x2(self, a: f64x2) -> m64x2

Checks if the elements in each lane of a are NaN.

pub fn is_not_nan_f32x4(self, a: f32x4) -> m32x4

Checks if the elements in each lane of a are not NaN.

pub fn is_not_nan_f64x2(self, a: f64x2) -> m64x2

Checks if the elements in each lane of a are not NaN.

pub fn max_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Computes the elementwise maximum of each lane of a and b.

pub fn max_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Computes the elementwise maximum of each lane of a and b.

pub fn min_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Computes the elementwise minimum of each lane of a and b.

pub fn min_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Computes the elementwise minimum of each lane of a and b.

pub fn mul_add_f32x4(self, a: f32x4, b: f32x4, c: f32x4) -> f32x4

Multiplies the elements of each lane of a and b and adds the result to c.

pub fn mul_add_f64x2(self, a: f64x2, b: f64x2, c: f64x2) -> f64x2

Multiplies the elements of each lane of a and b and adds the result to c.

pub fn mul_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Multiplies the elements of each lane of a and b.

pub fn mul_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Multiplies the elements of each lane of a and b.

pub fn not_i16x8(self, a: i16x8) -> i16x8

Returns the bitwise NOT of a.

pub fn not_i32x4(self, a: i32x4) -> i32x4

Returns the bitwise NOT of a.

pub fn not_i64x2(self, a: i64x2) -> i64x2

Returns the bitwise NOT of a.

pub fn not_i8x16(self, a: i8x16) -> i8x16

Returns the bitwise NOT of a.

pub fn not_m16x8(self, a: m16x8) -> m16x8

Returns the bitwise NOT of a.

pub fn not_m32x4(self, a: m32x4) -> m32x4

Returns the bitwise NOT of a.

pub fn not_m64x2(self, a: m64x2) -> m64x2

Returns the bitwise NOT of a.

pub fn not_m8x16(self, a: m8x16) -> m8x16

Returns the bitwise NOT of a.

pub fn not_u16x8(self, a: u16x8) -> u16x8

Returns the bitwise NOT of a.

pub fn not_u32x4(self, a: u32x4) -> u32x4

Returns the bitwise NOT of a.

pub fn not_u64x2(self, a: u64x2) -> u64x2

Returns the bitwise NOT of a.

pub fn not_u8x16(self, a: u8x16) -> u8x16

Returns the bitwise NOT of a.

pub fn or_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Returns the bitwise OR of a and b.

pub fn or_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Returns the bitwise OR of a and b.

pub fn or_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Returns the bitwise OR of a and b.

pub fn or_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Returns the bitwise OR of a and b.

pub fn or_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Returns the bitwise OR of a and b.

pub fn or_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Returns the bitwise OR of a and b.

pub fn or_m16x8(self, a: m16x8, b: m16x8) -> m16x8

Returns the bitwise OR of a and b.

pub fn or_m32x4(self, a: m32x4, b: m32x4) -> m32x4

Returns the bitwise OR of a and b.

pub fn or_m64x2(self, a: m64x2, b: m64x2) -> m64x2

Returns the bitwise OR of a and b.

pub fn or_m8x16(self, a: m8x16, b: m8x16) -> m8x16

Returns the bitwise OR of a and b.

pub fn or_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Returns the bitwise OR of a and b.

pub fn or_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Returns the bitwise OR of a and b.

pub fn or_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Returns the bitwise OR of a and b.

pub fn or_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Returns the bitwise OR of a and b.

pub fn reduce_max_f32x4(self, a: f32x4) -> f32

pub fn reduce_max_f64x2(self, a: f64x2) -> f64

pub fn reduce_min_f32x4(self, a: f32x4) -> f32

pub fn reduce_min_f64x2(self, a: f64x2) -> f64

pub fn reduce_product_f32x4(self, a: f32x4) -> f32

pub fn reduce_product_f64x2(self, a: f64x2) -> f64

pub fn reduce_sum_c32x2(self, a: f32x4) -> c32

pub fn reduce_sum_c64x1(self, a: f64x2) -> c64

pub fn reduce_sum_f32x4(self, a: f32x4) -> f32

pub fn reduce_sum_f64x2(self, a: f64x2) -> f64

pub fn select_f32x4(self, mask: m32x4, if_true: f32x4, if_false: f32x4) -> f32x4

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_f64x2(self, mask: m64x2, if_true: f64x2, if_false: f64x2) -> f64x2

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_i16x8(self, mask: m16x8, if_true: i16x8, if_false: i16x8) -> i16x8

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_i32x4(self, mask: m32x4, if_true: i32x4, if_false: i32x4) -> i32x4

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_i64x2(self, mask: m64x2, if_true: i64x2, if_false: i64x2) -> i64x2

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_i8x16(self, mask: m8x16, if_true: i8x16, if_false: i8x16) -> i8x16

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_u16x8(self, mask: m16x8, if_true: u16x8, if_false: u16x8) -> u16x8

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_u32x4(self, mask: m32x4, if_true: u32x4, if_false: u32x4) -> u32x4

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_u64x2(self, mask: m64x2, if_true: u64x2, if_false: u64x2) -> u64x2

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn select_u8x16(self, mask: m8x16, if_true: u8x16, if_false: u8x16) -> u8x16

Combines if_true and if_false, selecting elements from if_true if the corresponding mask in mask is set, otherwise selecting elements from if_false.

pub fn shl_const_i16x8<const AMOUNT: i32>(self, a: i16x8) -> i16x8

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_i32x4<const AMOUNT: i32>(self, a: i32x4) -> i32x4

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_i64x2<const AMOUNT: i32>(self, a: i64x2) -> i64x2

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_i8x16<const AMOUNT: i32>(self, a: i8x16) -> i8x16

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_u16x8<const AMOUNT: i32>(self, a: u16x8) -> u16x8

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_u32x4<const AMOUNT: i32>(self, a: u32x4) -> u32x4

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_u64x2<const AMOUNT: i32>(self, a: u64x2) -> u64x2

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_const_u8x16<const AMOUNT: i32>(self, a: u8x16) -> u8x16

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shl_dyn_i16x8(self, a: i16x8, amount: i16x8) -> i16x8

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_i32x4(self, a: i32x4, amount: i32x4) -> i32x4

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_i64x2(self, a: i64x2, amount: i64x2) -> i64x2

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_i8x16(self, a: i8x16, amount: i8x16) -> i8x16

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_u16x8(self, a: u16x8, amount: i16x8) -> u16x8

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_u32x4(self, a: u32x4, amount: i32x4) -> u32x4

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_u64x2(self, a: u64x2, amount: i64x2) -> u64x2

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shl_dyn_u8x16(self, a: u8x16, amount: i8x16) -> u8x16

Shift the bits of each lane of a to the left by the element in the corresponding lane in amount, while shifting in zeros.

pub fn shr_const_i16x8<const AMOUNT: i32>(self, a: i16x8) -> i16x8

Shift the bits of each lane of a to the left by AMOUNT, while shifting in sign bits.

pub fn shr_const_i32x4<const AMOUNT: i32>(self, a: i32x4) -> i32x4

Shift the bits of each lane of a to the left by AMOUNT, while shifting in sign bits.

pub fn shr_const_i64x2<const AMOUNT: i32>(self, a: i64x2) -> i64x2

Shift the bits of each lane of a to the left by AMOUNT, while shifting in sign bits.

pub fn shr_const_i8x16<const AMOUNT: i32>(self, a: i8x16) -> i8x16

Shift the bits of each lane of a to the left by AMOUNT, while shifting in sign bits.

pub fn shr_const_u16x8<const AMOUNT: i32>(self, a: u16x8) -> u16x8

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shr_const_u32x4<const AMOUNT: i32>(self, a: u32x4) -> u32x4

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shr_const_u64x2<const AMOUNT: i32>(self, a: u64x2) -> u64x2

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn shr_const_u8x16<const AMOUNT: i32>(self, a: u8x16) -> u8x16

Shift the bits of each lane of a to the left by AMOUNT, while shifting in zeros.

pub fn splat_f32x4(self, value: f32) -> f32x4

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_f64x2(self, value: f64) -> f64x2

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_i16x8(self, value: i16) -> i16x8

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_i32x4(self, value: i32) -> i32x4

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_i64x2(self, value: i64) -> i64x2

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_i8x16(self, value: i8) -> i8x16

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_m16x8(self, value: m16) -> m16x8

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_m32x4(self, value: m32) -> m32x4

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_m64x2(self, value: m64) -> m64x2

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_m8x16(self, value: m8) -> m8x16

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_u16x8(self, value: u16) -> u16x8

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_u32x4(self, value: u32) -> u32x4

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_u64x2(self, value: u64) -> u64x2

Returns a SIMD vector with all lanes set to the given value.

pub fn splat_u8x16(self, value: u8) -> u8x16

Returns a SIMD vector with all lanes set to the given value.

pub fn sub_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Subtracts the elements of each lane of a and b.

pub fn sub_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Subtracts the elements of each lane of a and b.

pub fn wrapping_add_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_add_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Adds the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn wrapping_sub_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Subtracts the elements of each lane of a and b, with wrapping on overflow.

pub fn xor_f32x4(self, a: f32x4, b: f32x4) -> f32x4

Returns the bitwise XOR of a and b.

pub fn xor_f64x2(self, a: f64x2, b: f64x2) -> f64x2

Returns the bitwise XOR of a and b.

pub fn xor_i16x8(self, a: i16x8, b: i16x8) -> i16x8

Returns the bitwise XOR of a and b.

pub fn xor_i32x4(self, a: i32x4, b: i32x4) -> i32x4

Returns the bitwise XOR of a and b.

pub fn xor_i64x2(self, a: i64x2, b: i64x2) -> i64x2

Returns the bitwise XOR of a and b.

pub fn xor_i8x16(self, a: i8x16, b: i8x16) -> i8x16

Returns the bitwise XOR of a and b.

pub fn xor_m16x8(self, a: m16x8, b: m16x8) -> m16x8

Returns the bitwise XOR of a and b.

pub fn xor_m32x4(self, a: m32x4, b: m32x4) -> m32x4

Returns the bitwise XOR of a and b.

pub fn xor_m64x2(self, a: m64x2, b: m64x2) -> m64x2

Returns the bitwise XOR of a and b.

pub fn xor_m8x16(self, a: m8x16, b: m8x16) -> m8x16

Returns the bitwise XOR of a and b.

pub fn xor_u16x8(self, a: u16x8, b: u16x8) -> u16x8

Returns the bitwise XOR of a and b.

pub fn xor_u32x4(self, a: u32x4, b: u32x4) -> u32x4

Returns the bitwise XOR of a and b.

pub fn xor_u64x2(self, a: u64x2, b: u64x2) -> u64x2

Returns the bitwise XOR of a and b.

pub fn xor_u8x16(self, a: u8x16, b: u8x16) -> u8x16

Returns the bitwise XOR of a and b.