pulp::x86

Struct V3Scalar

#[repr(transparent)]
pub struct V3Scalar {

    pub sse: Sse,
    pub sse2: Sse2,
    pub fxsr: Fxsr,
    pub sse3: Sse3,
    pub ssse3: Ssse3,
    pub sse4_1: Sse4_1,
    pub sse4_2: Sse4_2,
    pub popcnt: Popcnt,
    pub avx: Avx,
    pub avx2: Avx2,
    pub bmi1: Bmi1,
    pub bmi2: Bmi2,
    pub fma: Fma,
    pub lzcnt: Lzcnt,

}

AVX instruction set using only scalar instructions.

Notable additions over V2 include:

Instructions operating on 256-bit SIMD vectors.

Shift functions with a separate shift per lane, such as V3::shl_dyn_u32x4.

Fused multiply-accumulate instructions, such as V3::mul_add_f32x4.

Fields
sse: Ssesse2: Sse2fxsr: Fxsrsse3: Sse3ssse3: Ssse3sse4_1: Sse4_1sse4_2: Sse4_2popcnt: Popcntavx: Avxavx2: Avx2bmi1: Bmi1bmi2: Bmi2fma: Fmalzcnt: Lzcnt
Implementations

impl V3Scalar

AVX instruction set using only scalar instructions.

Notable additions over V2 include:

Instructions operating on 256-bit SIMD vectors.

Shift functions with a separate shift per lane, such as V3::shl_dyn_u32x4.

Fused multiply-accumulate instructions, such as V3::mul_add_f32x4.

pub unsafe fn new_unchecked() -> Self

Returns a SIMD token type without checking if the required CPU features for this type are available.

Safety

the required CPU features must be available.

pub fn try_new() -> Option<Self>

Returns a SIMD token type if the required CPU features for this type are available, otherwise returns None.

pub fn is_available() -> bool

Returns true if the required CPU features for this type are available, otherwise returns false.

pub fn vectorize<F: NullaryFnOnce>(self, f: F) -> F::Output

Vectorizes the given function as if the CPU features for this type were applied to it.

Note

For the vectorization to work properly, the given function must be inlined. Consider marking it as #[inline(always)]

pub fn to_ref(self) -> &'static Self

Takes a proof of the existence of this SIMD token (self), and returns a persistent reference to it.